rename : repository 모듈 하위에 있던, entity 모듈을 domain 모듈의 하위 모듈로 옮김.
