+++ moview/repository/interviewee_data_repository.py
@@ -2,10 +2,9 @@ from typing import Optional
 
 from pymongo import MongoClient
 
-from moview.repository.entity.interviewee_data_main_document import IntervieweeDataEntity
-from moview.repository.entity.interviewee_data_subdocument import *
+from moview.domain.entity.interviewee_data_main_document import IntervieweeDataEntity
 from moview.utils.singleton_meta_class import SingletonMeta
-
+from moview.domain.entity.interviewee_data_subdocument import *
 
 # todo 아직 몽고 db 연결전이므로 임시로 만들었음.
 class MongoConfig:
