+++ moview/service/interviewee_input/interviewee_input_service.py
@@ -3,8 +3,8 @@ from typing import Any
 from moview.modules.input.input_analyzer import InputAnalyzer
 from moview.modules.input.initial_question_giver import InitialQuestionGiver, InitialQuestionParseError
 from moview.repository.interviewee_data_repository import IntervieweeDataRepository, MongoConfig
-from moview.repository.entity.interviewee_data_main_document import IntervieweeDataEntity
-from moview.repository.entity.interviewee_data_subdocument import *
+from moview.domain.entity.interviewee_data_main_document import IntervieweeDataEntity
+from moview.domain.entity.interviewee_data_subdocument import *
 from moview.loggers.mongo_logger import *
