+++ tests/service/interviewee_answer/test_interviewee_answer_service.py
@@ -3,9 +3,9 @@ from unittest.mock import patch
 
 from moview.service.interviewee_answer.interviewer_action_enum import InterviewerActionEnum
 from moview.repository.interviewee_data_repository import IntervieweeDataRepository, MongoConfig
-from moview.repository.entity.interviewee_data_subdocument import *
-from moview.repository.entity.interviewee_data_main_document import IntervieweeDataEntity
+from moview.domain.entity.interviewee_data_main_document import IntervieweeDataEntity
 from moview.service.interviewee_answer.interviewee_answer_service import IntervieweeAnswerService
+from moview.domain.entity.interviewee_data_subdocument import *
 
 
 class TestAnswerServiceWithMocking(unittest.TestCase):
